pyproject.toml
[project]
name = "medicalbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.3",
    "ipykernel>=7.2.0",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.14",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.0",
    "langchain-pinecone>=0.2.13",
    "pypdf>=6.7.2",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.3",
]
